update checkpoint functions in layers
